pyproject.toml
[project]
name = "mar24cmlops-rakuten"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Jean Christophe CAZES", email = "cazes.jean.christophe@gmail.com" }
]
dependencies = [
    "fastapi>=0.111.0",
    "pandas>=2.2.2",
    "boto3>=1.34.117",
    "awscli>=1.32.117",
    "duckdb>=1.0.0",
    "cryptography>=42.0.8",
    "passlib>=1.7.4",
    "pyjwt>=2.8.0",
    "nltk>=3.8.1",
    "tensorflow>=2.16.2",
    "pillow>=10.4.0",
    "scikit-learn>=1.5.1",
    "torch>=2.3.1",
    "torchvision>=0.18.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.2.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mar24cmlops_rakuten"]
